+++ packages/monaco-editor-wrapper/CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/
 
 - Improve naming and improve api usage [#31](https://github.com/TypeFox/monaco-components/pull/31)
 - createUrl now allows web socket urls without port and path [#30](https://github.com/TypeFox/monaco-components/pull/30)
-- Updated to `monaco-languageclient` `6.4.0` using `monaco-vscode-api` `1.81.0` and `monaco-editor` `0.41.0`
+- Updated to `monaco-languageclient` `6.4.1` using `monaco-vscode-api` `1.81.1` and `monaco-editor` `0.41.0`
 
 ## [2.1.1] - 2023-07-27